refactor: split in components
